Big move towards data-driven-ness
-Bit of a refactor
-Small hack in DarknecEngine.h and State.h for component fiddling
-Settings are now loaded from XML. The loading method might be moved to
a different specialised class or file for XML loading
-GameObjects can only be instantiated from files for the time being to
avoid NULL pointers.
-Removed debug logging with an owner name due to method/argument
conflicts where the format ends up as the owner name from variadic arg
being shifted over.
-Object file parser created. Documentation will come later.

+++ src/object/fast_tiny_obj_loader.cc
@@ -345,8 +345,8 @@ namespace ObjLoader {
 
 			//Illumination model
 			else if (!strcmp(lineTag, "illum")) {
-				float illumination;
-				fscanf_s(file, "%f\n", &illumination);
+				int illumination;
+				fscanf_s(file, "%i\n", &illumination);
 				material.illum = illumination;
 			}
 			//Dissolve / Transparency
